backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
torch==1.13.1+cpu
torchvision==0.14.1+cpu
typing-extensions==4.4.0
urllib3==1.26.14
zipp==3.12.0
